backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
supabase
yfinance
python-dotenv
pydantic